% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{add_element}
\alias{add_element}
\title{Assigns object to a report via a text path.}
\usage{
add_element(item, location)
}
\arguments{
\item{item}{any. The object you'd like to store. Commonly dataframes, urls to
images, etc, but could be anything.}

\item{location}{character. The location in the report you'd like the object 
to be stored in, as a text string.}
}
\description{
Takes the object, forms the string report$[[object_path]], and
then assigns the object to that location in the report. E.g. 
add_element(some_plot, 'plots$plot_a') will take the plot object and will 
perform `report$plots$plot_a <- some_plot`
}

