% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_functions.R
\name{add_metadata_glmnet}
\alias{add_metadata_glmnet}
\title{Add metadata}
\usage{
add_metadata_glmnet(s3path, var_col, desc_col, additions = NULL)
}
\arguments{
\item{s3path}{S3 path to the metadata file.}

\item{var_col}{The name of the column that specifies the variable. For joining.}

\item{desc_col}{The name of the column that specifies the description.}

\item{additions}{Any additional column names from the metadata file you wish to add.}
}
\value{
report$coef
}
\description{
Add a file containing metadata to coefficients. Report$coef
must already be populated.
}

