% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{append_report}
\alias{append_report}
\title{Add something to your report.}
\usage{
append_report(report_path, func_list)
}
\arguments{
\item{report_path}{The s3 path to the report.}

\item{func_list}{List of functions to add, formatted exactly like your
report list.}
}
\value{
Technically does not return anything. But report elements are globally
assigned, so the new elements will be in memory.
}
\description{
Rather than rebuilding the report every time, append_report lets
you add functions to a report that already exists on s3. Good for testing new 
functions, troubleshooting, or ad hoc requests.
}

