% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{build_report}
\alias{build_report}
\title{Build the report list.}
\usage{
build_report(report_path, .disable_tests = FALSE)
}
\arguments{
\item{report_path}{The path to your report list.}
}
\value{
Technically does not return anything. Build_report writes the report
objects using global assignment, so you'll have the report object in memory
after execution, even if a particular function fails.
}
\description{
Specify the path to the file containing the report list.
Build_report reads the list, turns it into your report object, and uploads it
to the s3 path you specify.
}
\examples{
build_report('path/to/my/file.R')
}

