% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{check_coverage}
\alias{check_coverage}
\title{Check that all reporting functions have a non-trivial test.}
\usage{
check_coverage()
}
\value{
NULL. Prints a message to the screen when functions are missing.
}
\description{
Check that all reporting functions have a non-trivial test.
}

