% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{combine_sets}
\alias{combine_sets}
\title{Combines multiple s3 paths into a single data set}
\usage{
combine_sets(sets, list = "scored_data")
}
\arguments{
\item{sets}{character. A vector of the sets to include.}

\item{list}{character. The name of the set list to draw from.}
}
\value{
A data set containing the rbound data, , with column 'set' to 
distinguish the data's origin.
}
\description{
Combines multiple s3 paths into a single data set
}

