% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_scoring.R
\name{get_auc}
\alias{get_auc}
\title{Calculate auc for a scored data set.}
\usage{
get_auc(data_name, fold = NA)
}
\arguments{
\item{data_name}{The name of the data set, defined in the report list.}

\item{fold}{An optional param to calculate AUC split by the named column. Useful
for calculating AUC on cross-validation sets.}
}
\value{
report$auc$`set_name`
}
\description{
Specify one of the scored data sets. get_auc will calculate the
AUC for that set and append it to a vector of auc values.
}

