% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_scoring.R
\name{get_aucs}
\alias{get_aucs}
\title{Calculate AUCs for all scored sets.}
\usage{
get_aucs(data_names, folds = NA, location = "auc$val")
}
\arguments{
\item{folds}{An optional vector of the names of fold columns. Defaults to NA,
and will recycle. #'}

\item{data_paths}{A vector of the scored set names}
}
\value{
report$auc
}
\description{
Just sapply get_auc across every scored_set.
}

