% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_functions.R
\name{get_coef_glmnet}
\alias{get_coef_glmnet}
\title{Get the coefficients from a glmnet model.}
\usage{
get_coef_glmnet(s)
}
\arguments{
\item{s}{The lambda value for the coefficients.}
}
\value{
report$coef$`lambda_val`
}
\description{
Turns the ugly matrix presentation of model coefficients into
a nice, neat data frame.
}

