% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{get_element}
\alias{get_element}
\title{Retrieves object via a text path.}
\usage{
get_element(location)
}
\arguments{
\item{location}{character. The location in the report you'd like to retrieve
an object from.}
}
\description{
Works like add_element(), but retrieves the object at the 
specified text path.
}

