% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{lib}
\alias{lib}
\title{Helper to call the library environment from .syberiaReport, with all the
reporting functions}
\usage{
lib()
}
\value{
.syberiaReport$library
}
\description{
Helper to call the library environment from .syberiaReport, with all the
reporting functions
}

