% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_scoring.R
\name{make_decile_tables}
\alias{make_decile_tables}
\title{Adds decile info to report.}
\usage{
make_decile_tables(scored_sets, xcol = "score")
}
\arguments{
\item{scored_sets}{The names of the scored data sets.}

\item{xcol}{Which column to rank by decile. Defaults to score, but could be
used for other rank-orderings.}
}
\value{
report$performance$`colname_decile_table` and
report$performance$`colname_decile_plot`
}
\description{
Create a dataframe with set name, x column decile, and expected
and actual scores. Also appends a plot visualizing the differences between
expected and actual values.
}

