% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{make_report_env}
\alias{make_report_env}
\title{Create new, blank .ReportEnv}
\usage{
make_report_env(.disable_tests = FALSE)
}
\arguments{
\item{.disable_tests}{bool. Whether to run tests from syberiaReports.test
directory.}
}
\description{
Cleans the slate and creates a new reporting environment.
}

