% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testing.R
\name{perform_tests}
\alias{perform_tests}
\title{Executes all tests in the test library.}
\usage{
perform_tests()
}
\value{
NULL. Will print a series of warnings if functions fail tests.
}
\description{
Executes all tests in the test library.
}

