% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{store_plot}
\alias{store_plot}
\title{Write out a plot to s3}
\usage{
store_plot(plot_obj, name, plot_func = plot, opts = list())
}
\arguments{
\item{plot_obj}{The output from a plot function like ggplot() or plot()}

\item{name}{Give the plot a name. A cool one. No, cooler.}

\item{opts}{Optional parameters to pass as a list to s3_plot}
}
\value{
Returns the s3 path to the plot
}
\description{
. Really just a wrapper around s3_plot, but helps by managing the
plot paths for you.
}

