% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{tests}
\alias{tests}
\title{Helper to call the test environment from .syberiaReport, with all the
reporting functions}
\usage{
tests()
}
\value{
.syberiaReport$tests
}
\description{
Helper to call the test environment from .syberiaReport, with all the
reporting functions
}

