% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{get_set}
\alias{get_set}
\title{Get the s3 path to a set by using the report_list name.}
\usage{
get_set(name, list = "scored_data")
}
\arguments{
\item{name}{The name of the data set.}

\item{list}{Optional param to specify whether you want to use the scored data
list or the raw data list. Defaults to scored data.}
}
\value{
The s3link to the data set.
}
\description{
Gets a data set from the report_list section you specify.
}

